refactor: created data for DevTeam
